Make Selectable template covariant
With this, we express that one cannot use matching() to alter the
templated type of a Selectable, which means that it is OK to pass a
Selectable of a type to a method that accepts a Selectable of a parent
type.

The drawback is that it is no longer possible to use methods defined in
Collection on an object that is returned by matching().