Document recommended changes
We have done the upgrade of types in a 2 step process.
To take the example of the ORM, we should be able to allow
doctrine/collections 2 on the stable branch, and then add both parameter
and return type declarations on the unstable branch in one go. For that
reason, it is a bit more helpful to document the recommended changes
that should be done rather than what changed in 2.0.x